+++ modules/cms/classes/CmsObjectCollection.php
@@ -1,0 +1,14 @@
+<?php namespace Cms\Classes;
+
+use Illuminate\Support\Collection as CollectionBase;
+use System\Classes\ApplicationException;
+
+/**
+ * This class represents a collection of Cms Objects.
+ *
+ * @package october\cms
+ * @author Alexey Bobkov, Samuel Georges
+ */
+class CmsObjectCollection extends CollectionBase
+{
+}